Move the notary errors HTTP errors into the server package.
