+++ server/server.go
@@ -105,7 +105,8 @@ func RootHandler(ac auth.AccessController, ctx context.Context, trust signed.Cry
 
 	r.Methods("GET").Path("/_notary_server/health").HandlerFunc(health.StatusHandler)
 	r.Methods("GET").Path("/_notary_server/metrics").Handler(prometheus.Handler())
-	r.Methods("GET", "POST", "PUT", "HEAD", "DELETE").Path("/{other:.*}").Handler(hand(utils.NotFoundHandler))
+	r.Methods("GET", "POST", "PUT", "HEAD", "DELETE").Path("/{other:.*}").Handler(
+		hand(handlers.NotFoundHandler))
 
 	return r
 }
