+++ utils/http.go
@@ -8,7 +8,6 @@ import (
 	"github.com/docker/distribution/registry/api/errcode"
 	"github.com/docker/distribution/registry/api/v2"
 	"github.com/docker/distribution/registry/auth"
-	"github.com/docker/notary/errors"
 	"github.com/docker/notary/tuf/signed"
 	"github.com/gorilla/mux"
 	"golang.org/x/net/context"
@@ -94,10 +93,4 @@ func buildAccessRecords(repo string, actions ...string) []auth.Access {
 		})
 	}
 	return requiredAccess
-}
-
-// NotFoundHandler is used as a generic catch all handler to return the ErrMetadataNotFound
-// 404 response
-func NotFoundHandler(ctx context.Context, w http.ResponseWriter, r *http.Request) error {
-	return errors.ErrMetadataNotFound.WithDetail(nil)
 }
